+++ src/domain/enums/challenge.enum.ts
