+++ src/infrastructure/repositories/prisma/index.ts
@@ -1,0 +1,2 @@
+export { PrismaAuthRepository } from './auth.repository';
+export { PrismaUserRepository } from './user.repository';
