+++ src/infrastructure/services/index.ts
@@ -1,0 +1,2 @@
+export { AuthService } from './auth.service';
+export { UserService } from './user.service';
