+++ src/presentation/routes/index.ts
@@ -1,5 +1,5 @@
 import { Router } from 'express';
-import { AuthRoutes } from '../auth/presentation/routes';
+import { AuthRoutes } from './auth.routes';
 
 export class AppRoutes {
   static get routes(): Router {
